Always use the first row to set field names (#798)
* Always use the first row to set field names
* Disable QComboBox scrolling in FieldsForm
* Check metadata file exists before deleting

+++ ode/file.py
@@ -140,7 +140,8 @@ class File:
         """
         if self.path.is_file():
             self.path.unlink()
-            self.metadata_path.unlink()
+            if self.metadata_path.exists():
+                self.metadata_path.unlink()
         elif self.path.is_dir():
             shutil.rmtree(self.path)
             if self.metadata_path.exists():
